sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "molecule-sdk"
version = "0.2.0"
description = "Molecule AI SDK — build plugins (molecule_plugin) AND remote agents that join a Molecule AI org from another machine (molecule_agent)."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Molecule AI" }]
keywords = ["agents", "ai", "multi-agent", "a2a", "plugins", "saas", "remote-agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.31",
]

[project.urls]
Homepage = "https://github.com/hongmingw/molecule-monorepo"
Repository = "https://github.com/hongmingw/molecule-monorepo"
Documentation = "https://github.com/hongmingw/molecule-monorepo/tree/main/sdk/python"

[tool.setuptools.packages.find]
include = ["molecule_plugin*", "molecule_agent*"]
exclude = ["template*", "tests*"]
